+++ src/API/Def.js
@@ -4,7 +4,7 @@
  * Email: alexbuzin88@gmail.com
 */
 
-// [x]FIXME: Modify def for third parameter.
+// [x]#FIXME:10 Modify def for third parameter.
 /**
  * Defines variable. Makes convexPolyhedron object *CANNON.JS* from *THREE.JS* firure.
  *